% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rldbrda.R
\name{combine_data}
\alias{combine_data}
\title{Combine Redundant and Non-redundant Effect Sizes}
\usage{
combine_data(r2, cumul)
}
\arguments{
\item{r2}{A dataframe that is the output from `get_r2`, representing the
R-squared values for individual features.}

\item{cumul}{A dataframe that is the output from `get_cumul`, representing
the cumulative R-squared values from stepwise distance-based redundancy analysis.}
}
\value{
A dataframe containing the combined set of effect sizes, with rows
   representing variables and ordered by the R-squared values (`r2`) in
   descending order, and then by the adjusted cumulative R-squared values
   (`RDAcumul_R2.adj`). Column `Row.names` denotes the names of the variables.
}
\description{
This function takes the results from the `get_r2` function (representing redundant
effect sizes based on individual metadata variables) and the results from
the `get_cumul` function (representing non-redundant, cumulative effect sizes from
a stepwise db-RDA) and combines them into a single dataframe. The resulting dataframe
is ordered by redundant effect sizes (R-squared) in decreasing order and by
cumulative adjusted R-squared values.
}
