% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rldbrda.R
\name{get_cumul}
\alias{get_cumul}
\title{Perform Cumulative db-RDA Analysis}
\usage{
get_cumul(distmat, meta)
}
\arguments{
\item{distmat}{A distance matrix where rows and columns correspond to samples.}

\item{meta}{A dataframe containing metadata for each sample in `distmat`.}
}
\value{
A dataframe with the R-squared values for each variable added sequentially
    into the ordination model. The dataframe has columns adjusted for better
    naming convention and includes a cumulative R-squared value column and
    step count for each variable, along with the number of non-missing samples
    'RDAcumul_N'.
}
\description{
This function performs forward stepwise distance-based redundancy analysis (db-RDA)
to determine the cumulative effects of metadata variables on a distance matrix.
It utilizes the `capscale` and `ordiR2step` functions from the vegan package.
}
