% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rldbrda.R
\name{get_r2}
\alias{get_r2}
\title{Calculate R-squared Values for All Features Across Distance Matrix}
\usage{
get_r2(distmat, meta)
}
\arguments{
\item{distmat}{A distance matrix containing distances between sampled entities.}

\item{meta}{A dataframe containing metadata associated with `distmat`.}
}
\value{
A dataframe with rows for each feature and columns for the feature name,
    the F statistic 'Fa', the R-squared 'r2', the adjusted R-squared 'r2adj',
    the number of non-missing values 'N', the raw p-values 'pval', and the
    Benjamini-Hochberg adjusted p-values 'padj'. Row names of the dataframe
    are set to the feature names.
}
\description{
This function applies `get_r2_single` to each feature in the metadata dataframe
and calculates the R-squared and adjusted R-squared using constrained ordination
and permutation tests for all features. It also adjusts p-values using the
Benjamini-Hochberg method.
}
