% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rldbrda.R
\name{get_r2_single}
\alias{get_r2_single}
\title{Calculate R-squared Values for a Single Feature}
\usage{
get_r2_single(distmat, meta, feature)
}
\arguments{
\item{distmat}{A distance matrix in which distances between items are stored.}

\item{meta}{A data frame containing metadata associated with the items in `distmat`.}

\item{feature}{A character string specifying the column name in `meta` for which the
R-squared values are to be calculated.}
}
\value{
A data frame row with columns for the feature name, the F statistic 'Fa',
    the R-squared 'r2', the adjusted R-squared 'r2adj', the number of non-missing
    values 'N', and the associated p-value 'pval'.
}
\description{
This function computes the R-squared and adjusted R-squared for a single feature
from a distance matrix and its corresponding metadata using constrained
ordination (capscale) and permutation tests (anova.cca) from the 'vegan' package.
}
