% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rldbrda.R
\name{plot_dbrda}
\alias{plot_dbrda}
\title{Visualize db-RDA Effect Sizes}
\usage{
plot_dbrda(plot_data)
}
\arguments{
\item{plot_data}{A long-form dataframe prepared by `prepare_plot_data` containing
the variables with their effect sizes and an indicator for significance.
Expected columns are 'rowname' for features, 'variable' for effect type,
'value' for effect size, and 'significant' for significance indication.}
}
\value{
A ggplot object representing the horizontal bar plot of effect sizes.
  Bars are filled differently for R-squared and cumulative R-squared values
  and semi-transparent if the feature is not significant.
}
\description{
This function creates a horizontal bar plot to visualize the effect sizes from
db-RDA analysis. Features (variables) are displayed on the y-axis and their
effect sizes on the x-axis. The plot differentiates between individual adjusted
R-squared values and cumulative R-squared values, and indicates the significance
of features.
}
\examples{
library(RLdbRDA)
library(vegan)

data(varespec)
data(varechem)

out <- rldbrda(varespec, varechem)
out

plot_data <- prepare_plot_data(out)
plot_data

g <- plot_dbrda(plot_data)
g

}
