% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rldbrda.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Prepare Data for Plotting db-RDA Results}
\usage{
prepare_plot_data(dbrda_data)
}
\arguments{
\item{dbrda_data}{A dataframe from the output of `rldbrda` function containing
R-squared and cumulative R-squared values for metadata variables.}
}
\value{
A tidied long-form dataframe suitable for plotting with `ggplot2`.
  Columns are as follows:
  - `rowname`: Factor indicating the metadata variables in the order of R-squared values.
  - `variable`: Character, either 'r2adj' for adjusted R-squared or 'RDAcumul_R2.adj'
    for cumulative adjusted R-squared.
  - `value`: Numeric, the R-squared value associated with `variable`.
  - `significant`: Logical, indicating whether the variable is significant based on
    the presence of NAs replaced with the maximum R2 value.
}
\description{
This function takes the combined db-RDA results and prepares them for visualization
by restructuring the data, imputing missing values, and adding aesthetics
to indicate the significance of features.
}
\examples{
library(RLdbRDA)
library(vegan)

data(varespec)
data(varechem)

out <- rldbrda(varespec, varechem)
out

plot_data <- prepare_plot_data(out)
plot_data

g <- plot_dbrda(plot_data)
g

}
