% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rldbrda.R
\name{rldbrda}
\alias{rldbrda}
\title{RLdbRDA}
\usage{
rldbrda(df, meta, method = "bray", p_cutoff = 0.05)
}
\arguments{
\item{df}{The input data}

\item{meta}{The corresponding metadata}
}
\value{
a dataframe with the output
}
\description{
Do dbRDA analysis of a dataset and metadata
}
\examples{
library(RLdbRDA)
}
