% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rldbrda.R
\name{validate_input}
\alias{validate_input}
\title{Validate Input Data}
\usage{
validate_input(df, meta)
}
\arguments{
\item{df}{A data frame containing the data to be validated.}

\item{meta}{A data frame containing metadata associated with the data. Both `df` and
`meta` must have identical row names.}
}
\value{
None
}
\description{
This internal function checks if the input data frame matches the corresponding metadata
in terms of row names and also verifies whether the metadata contains any missing
values.
}
